+++ BWidgets/HRangeScrollBar.hpp
@@ -28,9 +28,8 @@
 #include "Supports/Scrollable.hpp"
 #include "Supports/KeyPressable.hpp"
 #include "../BEvents/WheelEvent.hpp"
+#include "../BEvents/KeyEvent.hpp"
 #include BWIDGETS_DEFAULT_DRAWKNOB_PATH
-#include <cairo/cairo.h>
-#include <utility>
 
 #ifndef BWIDGETS_DEFAULT_HRANGESCROLLBAR_WIDTH
 #define BWIDGETS_DEFAULT_HRANGESCROLLBAR_WIDTH BWIDGETS_DEFAULT_HSCROLLBAR_WIDTH
@@ -62,6 +61,8 @@ class HRangeScrollBar :	public Widget,
 						public KeyPressable
 {
 protected:
+
+	bool fineTuned_;
 
 public:
 	HScrollBar scrollbar;
@@ -175,6 +176,15 @@ public:
 	 *  Copies all properties from another %HRangeScrollBar. But NOT its linkage.
 	 */
 	void copy (const HRangeScrollBar* that);
+
+	/**
+     *  @brief  Sets the range step size.
+     *  @param step  Step size.
+	 *
+	 *  Also sets the number of sub steps to BWIDGETS_DEFAULT_NR_SUBSTEPS if 
+	 *  step size is 0.0.
+     */
+    virtual void setStep (const value_type& step) override;
 	
 	/**
      *  @brief  Optimizes the widget extends.
@@ -212,6 +222,26 @@ public:
 	 *  widget static callback function.
      */
     virtual void onWheelScrolled (BEvents::Event* event) override;
+
+	/**
+     *  @brief  Method when a KeyEvent with the type keyPressEvent is 
+     *  received.
+     *  @param event  Passed Event.
+     *
+     *  Overridable method called from the main window event scheduler if a
+     *  key is pressed. By default, it calls its static callback function.
+     */
+    virtual void onKeyPressed (BEvents::Event* event) override;
+
+    /**
+     *  @brief  Method when a KeyEvent with the type keyReleaseEvent is 
+     *  received.
+     *  @param event  Passed Event.
+     *
+     *  Overridable method called from the main window event scheduler if a
+     *  key is released. By default, it calls its static callback function.
+     */
+    virtual void onKeyReleased (BEvents::Event* event) override;
 
 	/**
      *  @brief  Sets the transfer function.
@@ -362,14 +392,18 @@ inline HRangeScrollBar::HRangeScrollBar	(const double  x, const double y, const 
 	ValueTransferable<value_type> ([transferFunc] (const value_type& x) {return value_type (transferFunc (x.first), transferFunc (x.second));},
 	 							   [reTransferFunc] (const value_type& x) {return value_type (reTransferFunc (x.first), reTransferFunc (x.second));}),
 	Scrollable (),
-	KeyPressable(),
+	KeyPressable(), 
+	fineTuned_(false),
 	scrollbar (urid, title),
 	button1 (urid, title),
 	button2 (urid, title),
 	symbol1 (Symbol::SymbolType::minus, urid, title),
 	symbol2 (Symbol::SymbolType::add, urid, title)
 {
-	setKeyPressable(false);	// Not supported yet
+	setKeyPressable(true);
+	grabDevice(BDevices::Keys(BDevices::Keys::KeyType::shiftL));
+	grabDevice(BDevices::Keys(BDevices::Keys::KeyType::shiftR));
+	if (step == 0.0) setNrSubs(BWIDGETS_DEFAULT_NR_SUBSTEPS);
 	setFocusText([](const Widget* w) {return	w->getTitle() + 
 												": " + 
 												(dynamic_cast<const HRangeScrollBar*>(w) ? 
@@ -413,12 +447,19 @@ inline void HRangeScrollBar::copy (const HRangeScrollBar* that)
 	button2.copy (&that->button2);
 	symbol1.copy (&that->symbol1);
 	symbol2.copy (&that->symbol2);
+	fineTuned_ = that->fineTuned_;
 	KeyPressable::operator=(*that);
 	Scrollable::operator= (*that);
 	ValueTransferable<value_type>::operator= (*that);
 	ValidatableRange<value_type>::operator= (*that);
 	ValueableTyped<value_type>::operator= (*that);
 	Widget::copy (that);
+}
+
+inline void HRangeScrollBar::setStep (const value_type& step)
+{
+	ValidatableRange<value_type>::setStep(step);
+	if ((step.first == 0.0) || (step.second == 0.0)) setNrSubs(BWIDGETS_DEFAULT_NR_SUBSTEPS);
 }
 
 inline void HRangeScrollBar::resize ()
@@ -494,16 +535,36 @@ inline void HRangeScrollBar::onWheelScrolled (BEvents::Event* event)
 	if (getEffectiveWidth() >= 1) 
 	{
 		value_type v = getValue();
-		if (getStep().first != 0.0) v.first += wev->getDelta().y * getStep ().first;
-		else v.first = getValueFromRatio (value_type (getRatioFromValue (v).first + wev->getDelta().y / getEffectiveWidth(), v.second)).first;
+		const double step = (fineTuned_ ?	1.0 / ((static_cast<double>(getNrSubs() + 1.0)) * getEffectiveWidth()) :
+											1.0 / getEffectiveWidth());
+		if (getStep().first != 0.0) v.first += wev->getDelta().y * (fineTuned_ ? getSubStep ().first : getStep ().first);
+		else v.first = getValueFromRatio (value_type (getRatioFromValue (v).first + wev->getDelta().y * step, v.second)).first;
 
-		if (getStep().second != 0.0) v.second -= wev->getDelta().y * getStep ().second;
-		else v.second = getValueFromRatio (value_type (v.first, getRatioFromValue (v).second - wev->getDelta().y / getEffectiveWidth())).second;
+		if (getStep().second != 0.0) v.second -= wev->getDelta().y * (fineTuned_ ? getSubStep ().second : getStep ().second);
+		else v.second = getValueFromRatio (value_type (v.first, getRatioFromValue (v).second - wev->getDelta().y * step)).second;
 
 		const value_type vr = getRatioFromValue (v);
 		if (vr.first <= vr.second) setValue (v);
 	}
 	Scrollable::onWheelScrolled (event);
+}
+
+inline void HRangeScrollBar::onKeyPressed (BEvents::Event* event)
+{
+	BEvents::KeyEvent* kev = dynamic_cast<BEvents::KeyEvent*>(event);
+	if (!kev) return;
+	if (kev->getWidget() == this) fineTuned_ = true;
+	
+	KeyPressable::onKeyPressed(event);
+}
+
+inline void HRangeScrollBar::onKeyReleased (BEvents::Event* event)
+{
+	BEvents::KeyEvent* kev = dynamic_cast<BEvents::KeyEvent*>(event);
+	if (!kev) return;
+	if (kev->getWidget() == this) fineTuned_ = false;
+
+	KeyPressable::onKeyReleased(event);
 }
 
 inline void HRangeScrollBar::setTransferFunction (std::function<value_type (const value_type& x)> func)