+++ BWidgets/HScale.hpp
@@ -24,6 +24,7 @@
 #include "Supports/Scrollable.hpp"
 #include "Supports/KeyPressable.hpp"
 #include "../BEvents/WheelEvent.hpp"
+#include "../BEvents/KeyEvent.hpp"
 #include BWIDGETS_DEFAULT_DRAWHBAR_PATH
 
 #ifndef BWIDGETS_DEFAULT_HSCALE_WIDTH
@@ -51,6 +52,11 @@ class HScale :	public HMeter,
 				public Scrollable,
 				public KeyPressable
 {
+
+protected:
+
+	bool fineTuned_;
+
 public:
 
 	/**
@@ -121,6 +127,15 @@ public:
 	void copy (const HScale* that);
 
 	/**
+     *  @brief  Sets the range step size.
+     *  @param step  Step size.
+	 *
+	 *  Also sets the number of sub steps to BWIDGETS_DEFAULT_NR_SUBSTEPS if 
+	 *  step size is 0.0.
+     */
+    virtual void setStep (const double& step) override;
+
+	/**
      *  @brief  Method called when pointer button pressed.
      *  @param event  Passed Event.
      *
@@ -149,6 +164,26 @@ public:
 	 *  widget static callback function.
      */
     virtual void onWheelScrolled (BEvents::Event* event) override;
+
+	/**
+     *  @brief  Method when a KeyEvent with the type keyPressEvent is 
+     *  received.
+     *  @param event  Passed Event.
+     *
+     *  Overridable method called from the main window event scheduler if a
+     *  key is pressed. By default, it calls its static callback function.
+     */
+    virtual void onKeyPressed (BEvents::Event* event) override;
+
+    /**
+     *  @brief  Method when a KeyEvent with the type keyReleaseEvent is 
+     *  received.
+     *  @param event  Passed Event.
+     *
+     *  Overridable method called from the main window event scheduler if a
+     *  key is released. By default, it calls its static callback function.
+     */
+    virtual void onKeyReleased (BEvents::Event* event) override;
 
 protected:
 	/**
@@ -208,9 +243,13 @@ inline HScale::HScale	(const double  x, const double y, const double width, cons
 		Clickable(),
 		Draggable(),
 		Scrollable(),
-		KeyPressable()
+		KeyPressable(), 
+		fineTuned_(false)
 {
-	setKeyPressable(false);	// Not supported yet
+	setKeyPressable(true);
+	grabDevice(BDevices::Keys(BDevices::Keys::KeyType::shiftL));
+	grabDevice(BDevices::Keys(BDevices::Keys::KeyType::shiftR));
+	if (step == 0.0) setNrSubs(BWIDGETS_DEFAULT_NR_SUBSTEPS);
 }
 
 inline Widget* HScale::clone () const 
@@ -222,12 +261,18 @@ inline Widget* HScale::clone () const
 
 inline void HScale::copy (const HScale* that)
 {
-	scale_ = that->scale_;
+	fineTuned_ = that->fineTuned_;
 	KeyPressable::operator=(*that);
 	Scrollable::operator= (*that);
 	Draggable::operator= (*that);
 	Clickable::operator= (*that);
 	HMeter::copy (that);
+}
+
+inline void HScale::setStep (const double &step)
+{
+	ValidatableRange<double>::setStep(step);
+	if (step == 0.0) setNrSubs(BWIDGETS_DEFAULT_NR_SUBSTEPS);
 }
 
 inline void HScale::onButtonPressed (BEvents::Event* event)
@@ -251,8 +296,13 @@ inline void HScale::onPointerDragged (BEvents::Event* event)
 		if (!pev) return;
 		if (scale_.getWidth()) 
 		{
-			if (getStep() != 0.0) setValue (getValue() - pev->getDelta().y * getStep ());
-			else setValue (getValueFromRatio (getRatioFromValue(getValue()) - pev->getDelta().y / scale_.getWidth()));
+			if (getStep() != 0.0) setValue (getValue() - pev->getDelta().y * (fineTuned_ ?	getSubStep() : getStep()));
+			else 
+			{
+				const double step = (fineTuned_ ?	1.0 / ((static_cast<double>(getNrSubs() + 1.0)) * scale_.getWidth()) :
+													1.0 / scale_.getWidth());
+				setValue (getValueFromRatio	(getRatioFromValue (getValue()) - pev->getDelta().y * step));
+			}
 		}
 	}
 	Draggable::onPointerDragged (event);
@@ -264,10 +314,33 @@ inline void HScale::onWheelScrolled (BEvents::Event* event)
 	if (!wev) return;
 	if (scale_.getWidth()) 
 	{
-		if (getStep() != 0.0) setValue (getValue() - wev->getDelta().y * getStep ());
-		else setValue (getValueFromRatio (getRatioFromValue(getValue()) - wev->getDelta().y / scale_.getWidth()));
+		if (getStep() != 0.0) setValue (getValue() - wev->getDelta().y * (fineTuned_ ?	getSubStep() : getStep()));
+		else 
+		{
+			const double step = (fineTuned_ ?	1.0 / ((static_cast<double>(getNrSubs() + 1.0)) * scale_.getWidth()) :
+												1.0 / scale_.getWidth());
+			setValue (getValueFromRatio	(getRatioFromValue (getValue()) - wev->getDelta().y * step));
+		}
 	}
 	Scrollable::onWheelScrolled (event);
+}
+
+inline void HScale::onKeyPressed (BEvents::Event* event)
+{
+	BEvents::KeyEvent* kev = dynamic_cast<BEvents::KeyEvent*>(event);
+	if (!kev) return;
+	if (kev->getWidget() == this) fineTuned_ = true;
+	
+	KeyPressable::onKeyPressed(event);
+}
+
+inline void HScale::onKeyReleased (BEvents::Event* event)
+{
+	BEvents::KeyEvent* kev = dynamic_cast<BEvents::KeyEvent*>(event);
+	if (!kev) return;
+	if (kev->getWidget() == this) fineTuned_ = false;
+
+	KeyPressable::onKeyReleased(event);
 }
 
 inline void HScale::draw ()